requirements.txt
Django==3.1.2
psycopg2-binary==2.8.6
gunicorn==20.0.4

djangorestframework==3.12.1
markdown==3.3.2
django-filter==2.4.0
django-crispy-forms==1.9.2